Finish button input implementation
