+++ m3dialib/includes/m3d/core/core.hpp
@@ -13,6 +13,7 @@
 #include "ledPattern.hpp"
 #include "thread.hpp"
 #include "time.hpp"
+#include "input.hpp"
 
 
 #endif /* end of include guard: CORE_H */
