+++ m3dialib/includes/m3d/input/input.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "buttons.hpp"
+// #include "buttons.hpp"
 #include "circlepad.hpp"
 #include "softwareKeyboard.hpp"
 #include "touch.hpp"
