+++ m3dialib/source/core/input.cpp
@@ -1,7 +1,6 @@
 #include <3ds.h>
 #include <cstring>
 #include "m3d/core/input.hpp"
-#include "m3d/input/buttons.hpp"
 
 namespace m3d {
     u32 Input::buttonsDown() {
@@ -16,15 +15,15 @@ namespace m3d {
         return hidKeysUp();
     }
 
-    bool Input::buttonPressed(m3d::buttons::Button t_button) {
+    bool Input::buttonPressed(Button t_button) {
         return (buttonsPressed() & t_button);
     }
 
-    bool Input::buttonDown(m3d::buttons::Button t_button) {
+    bool Input::buttonDown(Button t_button) {
         return (buttonsDown() & t_button);
     }
 
-    bool Input::buttonReleased(m3d::buttons::Button t_button) {
+    bool Input::buttonReleased(Button t_button) {
         return (buttonsReleased() & t_button);
     }
 }
